% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdata.R
\name{bigdata}
\alias{bigdata}
\title{A function to report on my dog's weight}
\usage{
bigdata(dog, weight, day)
}
\arguments{
\item{dog}{The name of a dog}

\item{weight}{A weight in pounds of the dog}

\item{day}{day of the week the weight was taken on}
}
\description{
A function to report on my dog's weight
}
\examples{
bigdata("Cholla", 50, "Friday")
}
